requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
postgres==4.0
prometheus_client==0.21.1
prometheus_flask_exporter==0.23.2
psycopg2-binary==2.9.10
psycopg2-pool==1.2
python-dotenv==1.1.0
redis==5.2.1
SQLAlchemy==2.0.40
typing_extensions==4.13.0
Werkzeug==3.1.3
